Rename settings module and app module
