+++ .gitignore
@@ -1,6 +1,6 @@
-my_books/migrations/__init__.py
-my_books/migrations/0001_initial.py
-my_book/__pycache__/__init__.cpython-38.pyc
-my_book/__pycache__/settings.cpython-38.pyc
+books/migrations/__init__.py
+books/migrations/0001_initial.py
+core/__pycache__/__init__.cpython-38.pyc
+core/__pycache__/settings.cpython-38.pyc
 Pipfile
 Pipfile.lock
