+++ books/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MyBooksConfig(AppConfig):
-    name = 'my_books'
+    name = 'books'
