+++ core/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for my_book project.
+Django settings for core project.
 
 Generated by 'django-admin startproject' using Django 3.1.1.
 
@@ -46,12 +46,12 @@ INSTALLED_APPS = [
     'django_filters',
 
     # My apps
-    'my_books',
+    'books',
 ]
 
 REST_FRAMEWORK = {
     'DEFAULT_PAGINATION_CLASS':
-    'my_books.custompagination.LimitOffsetPaginationWithUpperBound',
+    'books.custompagination.LimitOffsetPaginationWithUpperBound',
     'PAGE_SIZE': 5,
     'DEFAULT_FILTER_BACKENDS': (
         'django_filters.rest_framework.DjangoFilterBackend',
@@ -64,7 +64,7 @@ REST_FRAMEWORK = {
         ),
     'DEFAULT_PERMISSION_CLASSES':(
         'rest_framework.permissions.IsAuthenticatedOrReadOnly',
-        'my_books.custompermission.IsAdminUserOrReadOnly',
+        'books.custompermission.IsAdminUserOrReadOnly',
         ),
     'DEFAULT_THROTTLE_CLASSES':(
         'rest_framework.throttling.AnonRateThrottle',
@@ -86,7 +86,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'my_book.urls'
+ROOT_URLCONF = 'core.urls'
 
 TEMPLATES = [
     {
@@ -104,7 +104,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'my_book.wsgi.application'
+WSGI_APPLICATION = 'core.wsgi.application'
 
 
 # Database
