+++ core/urls.py
@@ -4,7 +4,7 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     # url for api
-    path('api/v1/', include('my_books.urls')),
+    path('api/v1/', include('books.urls')),
     # url for authentication system
     path('api-auth/', include('rest_framework.urls')),
 ]
